+++ src/classes/filerenaming/handlers/XMLFileRenameHandler.ts
@@ -1,57 +1,61 @@
 import * as vscode from "vscode";
 import { FileReader } from "../../utils/FileReader";
-import { FileRenameHandler } from "./abstraction/FileRenameHandler";
+import { FileRenameHandler, IFileChanges } from "./abstraction/FileRenameHandler";
 import * as fs from "fs";
-import * as path from "path";
 import { DiagnosticsRegistrator } from "../../registrators/DiagnosticsRegistrator";
-const fileSeparator = path.sep;
 function escapeRegExp(string: string) {
 	return string.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
 }
 
 export class XMLFileRenameHandler extends FileRenameHandler {
-	handleFileRename(oldUri: vscode.Uri, newUri: vscode.Uri): void {
+	handleFileRename(oldUri: vscode.Uri, newUri: vscode.Uri, allFiles: IFileChanges[]): IFileChanges[] {
 		if (newUri.fsPath.endsWith(".view.xml")) {
-			this._replaceViewNames(oldUri, newUri);
+			this._replaceViewNames(oldUri, newUri, allFiles);
 		} else if (newUri.fsPath.endsWith(".fragment.xml")) {
-			this._replaceFragmentNames(oldUri, newUri);
+			this._replaceFragmentNames(oldUri, newUri, allFiles);
 		}
 
-		const oldClassName = FileReader.getClassNameFromPath(oldUri.fsPath);
-		const newClassName = FileReader.getClassNameFromPath(newUri.fsPath);
-		if (oldClassName && newClassName) {
-			if (newUri.fsPath.endsWith(".view.xml")) {
-				FileReader.replaceViewNames(oldClassName, newClassName);
-			} else if (newUri.fsPath.endsWith(".fragment.xml")) {
-				FileReader.replaceFragmentNames(oldClassName, newClassName);
+		DiagnosticsRegistrator.removeDiagnosticForUri(oldUri, "xml");
+		// const newFile = allFiles.find(file => file.fileData.fsPath === newUri.fsPath);
+		// if (newFile) {
+		// 	newFile.changed = true;
+		// }
+		const oldName = FileReader.getClassNameFromPath(oldUri.fsPath);
+		const newName = FileReader.getClassNameFromPath(newUri.fsPath);
+		if (oldName && newName) {
+			if (oldUri.fsPath.endsWith(".fragment.xml")) {
+				FileReader.replaceFragmentNames(oldName, newName);
+			} else if (oldUri.fsPath.endsWith(".view.xml")) {
+				FileReader.replaceViewNames(oldName, newName);
 			}
 		}
-		DiagnosticsRegistrator.removeDiagnosticForUri(oldUri, "xml");
+
+		return allFiles;
 	}
 
 
-	private _replaceFragmentNames(oldUri: vscode.Uri, newUri: vscode.Uri) {
+	private _replaceFragmentNames(oldUri: vscode.Uri, newUri: vscode.Uri, allFiles: IFileChanges[]) {
 		const textToReplaceFromDotNotation = FileReader.getClassNameFromPath(oldUri.fsPath)?.replace(".fragment.xml", "");
 		const textToReplaceToDotNotation = FileReader.getClassNameFromPath(newUri.fsPath)?.replace(".fragment.xml", "");
 
 		if (textToReplaceFromDotNotation && textToReplaceToDotNotation) {
-			this.replaceAllOccurrencesInFiles(textToReplaceFromDotNotation, textToReplaceToDotNotation);
+			this.replaceAllOccurrencesInFiles(textToReplaceFromDotNotation, textToReplaceToDotNotation, allFiles);
 		}
 	}
 
 
-	private _replaceViewNames(oldUri: vscode.Uri, newUri: vscode.Uri) {
+	private _replaceViewNames(oldUri: vscode.Uri, newUri: vscode.Uri, allFiles: IFileChanges[]) {
 		const textToReplaceFromDotNotation = FileReader.getClassNameFromPath(oldUri.fsPath)?.replace(".view.xml", "");
 		const textToReplaceToDotNotation = FileReader.getClassNameFromPath(newUri.fsPath)?.replace(".view.xml", "");
 
 		if (textToReplaceFromDotNotation && textToReplaceToDotNotation) {
-			this._renameController(textToReplaceToDotNotation);
-			this._replaceViewNamesInManifests(textToReplaceFromDotNotation, textToReplaceToDotNotation);
-			this.replaceAllOccurrencesInFiles(textToReplaceFromDotNotation, textToReplaceToDotNotation);
+			this._renameController(textToReplaceToDotNotation, allFiles);
+			this._replaceViewNamesInManifests(textToReplaceFromDotNotation, textToReplaceToDotNotation, allFiles);
+			this.replaceAllOccurrencesInFiles(textToReplaceFromDotNotation, textToReplaceToDotNotation, allFiles);
 		}
 	}
 
-	private _renameController(newViewName: string) {
+	private _renameController(newViewName: string, allFiles: IFileChanges[]) {
 		const viewNamePart = newViewName.split(".")[newViewName.split(".").length - 1];
 		const viewPath = FileReader.convertClassNameToFSPath(newViewName, false, false, true);
 		if (viewPath) {
@@ -64,36 +68,43 @@ export class XMLFileRenameHandler extends FileRenameHandler {
 					newControllerNameParts[newControllerNameParts.length - 1] = viewNamePart;
 					const newControllerName = newControllerNameParts.join(".");
 					const newControllerPath = FileReader.convertClassNameToFSPath(newControllerName, true);
-					if (newControllerPath) {
-						fs.renameSync(controllerPath, newControllerPath);
-						const oldUri = vscode.Uri.file(controllerPath);
-						const newUri = vscode.Uri.file(newControllerPath);
-						this.replaceCurrentClassNameWithNewOne(oldUri, newUri);
+					if (newControllerPath && controllerPath !== newControllerPath) {
+						const controllerFile = allFiles.find(file => file.fileData.fsPath === controllerPath);
+						if (controllerFile) {
+							controllerFile.renames.push({
+								oldFSPath: controllerPath,
+								newFSPath: newControllerPath
+							});
+						}
 					}
 				}
 			}
 		}
 	}
 
-	private _replaceViewNamesInManifests(textToReplaceFromDotNotation: string, textToReplaceToDotNotation: string) {
-		const manifests = FileReader.getAllManifests();
+	private _replaceViewNamesInManifests(textToReplaceFromDotNotation: string, textToReplaceToDotNotation: string, allFiles: IFileChanges[]) {
+		const manifests = allFiles.filter(file => file.fileData.fsPath.endsWith("manifest.json"));
 
 		manifests.forEach(manifest => {
-			const viewPath = manifest.content["sap.ui5"]?.routing?.config?.viewPath;
+			try {
+				const content = JSON.parse(manifest.fileData.content);
+				const viewPath = content["sap.ui5"]?.routing?.config?.viewPath;
 
-			if (viewPath && textToReplaceFromDotNotation.startsWith(viewPath)) {
-				const oldPath = `"${textToReplaceFromDotNotation.replace(viewPath, "").replace(".", "")}"`/*removes first dot*/;
-				const newPath = `"${textToReplaceToDotNotation.replace(viewPath, "").replace(".", "")}"`/*removes first dot*/;
+				if (viewPath && textToReplaceFromDotNotation.startsWith(viewPath)) {
+					const oldPath = `"${textToReplaceFromDotNotation.replace(viewPath, "").replace(".", "")}"`/*removes first dot*/;
+					const newPath = `"${textToReplaceToDotNotation.replace(viewPath, "").replace(".", "")}"`/*removes first dot*/;
 
-				if (JSON.stringify(manifest.content).indexOf(oldPath) > -1) {
-					const fsPath = `${manifest.fsPath}${fileSeparator}manifest.json`;
-					let manifestText = fs.readFileSync(fsPath, "utf8");
-					manifestText = manifestText.replace(new RegExp(`${escapeRegExp(oldPath)}`, "g"), newPath);
-					fs.writeFileSync(fsPath, manifestText);
+					if (JSON.stringify(content).indexOf(oldPath) > -1) {
+						// const fsPath = `${manifest.fsPath}${fileSeparator}manifest.json`;
+						// let manifestText = fs.readFileSync(fsPath, "utf8");
+						manifest.fileData.content = manifest.fileData.content.replace(new RegExp(`${escapeRegExp(oldPath)}`, "g"), newPath);
+						// fs.writeFileSync(fsPath, manifestText);
+						manifest.changed = true;
+					}
 				}
+			} catch (error) {
+				console.error(error.message);
 			}
 		});
-
-		FileReader.rereadAllManifests();
 	}
 }