+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.12.42 (19-05-2021)
+* Added error message for failed http requests
+* Bugfixes for file/folder rename handler
+
 ## 0.12.41 (17-05-2021)
 * JS Rename provider fix